flask_app/requirements.txt
Flask>=2.0
PyMySQL>=1.0
flask-cors>=3.0
python-dotenv>=1.0
pytest>=7.0
requests>=2.0
gunicorn>=20.1
